+++ com.ibm.streamsx.kafka/CHANGELOG.md
@@ -1,11 +1,21 @@
 # Changes
 ==========
+## latest/develop
+* [#220](https://github.com/IBMStreams/streamsx.kafka/issues/220) SPL doc: Add hints, what to do when the toolkit is used with Kafka version 0.10
+* [#218](https://github.com/IBMStreams/streamsx.kafka/issues/218) support incremental cooperative rebalancing
+* [#217](https://github.com/IBMStreams/streamsx.kafka/issues/217) upgrade kafka-clients from 2.3.1 to 2.5.1
+
 ## v3.1.3
 * [#223](https://github.com/IBMStreams/streamsx.kafka/issues/223) I18n: update message translation
 
 ## v3.1.2
 * [#224](https://github.com/IBMStreams/streamsx.kafka/issues/224) Application directory cannot be used as file location for files specified via system property and vmArg
 * [#222](https://github.com/IBMStreams/streamsx.kafka/issues/222) KafkaConsumer: Reset of CR may fail when partitions are revoked before the reset actually happens
+
+## v3.2.0
+* [#220](https://github.com/IBMStreams/streamsx.kafka/issues/220) SPL doc: Add hints, what to do when the toolkit is used with Kafka version 0.10
+* [#218](https://github.com/IBMStreams/streamsx.kafka/issues/218) support incremental cooperative rebalancing
+* [#217](https://github.com/IBMStreams/streamsx.kafka/issues/217) upgrade kafka-clients from 2.3.1 to 2.5.1, for more information, see also the [release page](https://github.com/IBMStreams/streamsx.kafka/releases/tag/v3.2.0)
 
 ## v3.1.1
 * [#216](https://github.com/IBMStreams/streamsx.kafka/issues/216) Toolkit includes a vulnerable version of log4j.